Refactors, typing updates
