+++ package.json
@@ -16,7 +16,7 @@
     "node": "10.16.3"
   },
   "dependencies": {
-    "@material-ui/core": "^4.0.0-alpha.5",
+    "@material-ui/core": "^4.6.1",
     "@material-ui/icons": "^4.4.3",
     "@types/d3": "^5.7.2",
     "autoprefixer": "^9.5.1",
