+++ src/reducers/datasets.ts
@@ -1,18 +1,17 @@
 import { IDataset } from '../types/'
 import { findMatchingEntities } from './'
 import { FETCH_DATASETS_SUCCESS, FIND_MATCHING_ENTITIES } from '../constants/ActionTypes'
+import {
+  fetchDatasetsSuccess,
+  findMatchingEntities as findMatchingEntitiesActionCreator
+} from '../actionCreators'
 
 export type IDatasetsState = IDataset[]
 
 export const initialState: IDatasetsState = []
 
-interface IDatasetsAction {
-  type: string
-  payload: {
-    datasets?: IDataset[]
-    search?: string
-  }
-}
+type IDatasetsAction = ReturnType<typeof fetchDatasetsSuccess> &
+  ReturnType<typeof findMatchingEntitiesActionCreator>
 
 export default (state: IDatasetsState = initialState, action: IDatasetsAction): IDatasetsState => {
   const { type, payload } = action
