+++ src/reducers/jobs.ts
@@ -1,18 +1,17 @@
 import { IJob } from '../types/'
 import { findMatchingEntities } from './'
 import { FETCH_JOBS_SUCCESS, FIND_MATCHING_ENTITIES } from '../constants/ActionTypes'
+import {
+  fetchJobsSuccess,
+  findMatchingEntities as findMatchingEntitiesActionCreator
+} from '../actionCreators'
 
 export type IJobsState = IJob[]
 
 export const initialState: IJobsState = []
 
-interface IJobsAction {
-  type: string
-  payload: {
-    jobs: IJob[]
-    search?: string
-  }
-}
+type IJobsAction = ReturnType<typeof fetchJobsSuccess> &
+  ReturnType<typeof findMatchingEntitiesActionCreator>
 
 export default (state = initialState, action: IJobsAction): IJobsState => {
   const { type, payload } = action
